# [quant] update FakeQuant modules to use tensor qparams