--- a/test/quantization/core/test_workflow_ops.py
+++ b/test/quantization/core/test_workflow_ops.py
@@ -96,7 +96,7 @@ def _fake_quantize_learnable_per_channel_affine_grad_reference(
     - https://arxiv.org/pdf/1902.08153.pdf
     - https://arxiv.org/pdf/1903.08066.pdf
     """
-    per_channel_zero_point = ((per_channel_zero_point.detach() + 0.5).clamp(quant_min, quant_max)).type(torch.int64)
+    per_channel_zero_point = ((per_channel_zero_point.detach() + 0.5).clamp(quant_min, quant_max)).type(torch.int32)
     grad_X = _fake_quantize_per_channel_affine_grad_reference(
         dY, X, per_channel_scale, per_channel_zero_point, axis, quant_min, quant_max).to(device)
     per_channel_scale = per_channel_scale.detach().type(torch.float)
@@ -309,7 +309,7 @@ class TestFakeQuantizeOps(TestCase):
 
             X = X_base.clone().float()
             scale_base = scale_base.to(device).float()
-            zero_point_base = zero_point_base.to(dtype=torch.int64, device=device)
+            zero_point_base = zero_point_base.to(dtype=torch.int32, device=device)
             scale = scale_base.clone()
             zero_point = zero_point_base.clamp(quant_min, quant_max)
 
@@ -445,6 +445,7 @@ class TestFakeQuantizeOps(TestCase):
         X, (scale, zero_point, torch_type) = X
         X = to_tensor(X, device)
         fq_module = default_affine_fixed_qparams_fake_quant()
+        fq_module.to(device)
         fixed_scale = fq_module.scale.clone()
         fixed_zero_point = fq_module.zero_point.clone()
         # run fq module and make sure the quantization parameters does not change
@@ -585,7 +586,7 @@ class TestFakeQuantizeOps(TestCase):
 
         X = to_tensor(X, device)
         scale = to_tensor(scale, device)
-        zero_point = torch.tensor(zero_point).to(dtype=torch.int64, device=device)
+        zero_point = torch.tensor(zero_point).to(dtype=torch.int32, device=device)
         Y = _fake_quantize_per_channel_affine_reference(X.cpu(), scale.cpu(), zero_point.cpu(), axis, quant_min, quant_max)
         Y_prime = torch.fake_quantize_per_channel_affine(
             X, scale, zero_point, axis, quant_min, quant_max)
@@ -602,7 +603,7 @@ class TestFakeQuantizeOps(TestCase):
             obs(X * 0.75)
             scale, zero_point = obs.calculate_qparams()
             # TODO(future PR): fix the wrong dtype in obs.calculate_qparams and remove the cast
-            zero_point = zero_point.to(torch.int64)
+            zero_point = zero_point.to(torch.int32)
             quant_min, quant_max = obs._calculate_qmin_qmax()
 
             Y = _fake_quantize_per_channel_affine_reference(
@@ -621,7 +622,7 @@ class TestFakeQuantizeOps(TestCase):
 
     def test_forward_per_channel_half_precision_numerics(self):
         scale = torch.randn(5).abs()
-        zero = torch.randn(5).to(dtype=torch.long)
+        zero = torch.randn(5).to(dtype=torch.int)
         axis = 1
         mini = 0
         maxi = 255
@@ -708,7 +709,7 @@ class TestFakeQuantizeOps(TestCase):
 
         X = to_tensor(X, device)
         scale = to_tensor(scale, device)
-        zero_point = torch.tensor(zero_point).to(dtype=torch.int64, device=device)
+        zero_point = torch.tensor(zero_point).to(dtype=torch.int32, device=device)
         X.requires_grad_()
         Y_prime = torch.fake_quantize_per_channel_affine(
             X, scale, zero_point, axis, quant_min, quant_max)
@@ -729,7 +730,7 @@ class TestFakeQuantizeOps(TestCase):
             obs(X * 0.75)
             scale, zero_point = obs.calculate_qparams()
             # TODO(future PR): fix the wrong dtype in obs.calculate_qparams and remove the cast
-            zero_point = zero_point.to(torch.int64)
+            zero_point = zero_point.to(torch.int32)
             quant_min, quant_max = obs._calculate_qmin_qmax()
             X.requires_grad_()
             Y_prime = torch.fake_quantize_per_channel_affine(
@@ -837,7 +838,7 @@ class TestFakeQuantizeOps(TestCase):
         torch.random.manual_seed(NP_RANDOM_SEED)
         torch_types = [torch.qint8, torch.quint8]
         float_types = [torch.float, torch.float16, torch.float64]
-        zero_types = [torch.long]
+        zero_types = [torch.int]
         devices = [torch.device('cpu'), torch.device('cuda')] if torch.cuda.is_available() else [torch.device('cpu')]
         axis = 1
         for i in range(20):